reduced all data related modules into the data manager module, and merged the paths handler class into subject class. we are now able to properly run notebooks in colab since we don't have dependencies between modules in the package. added a requirements file. added support for gpu use in model training. added a new notebook for experimenting. adjusted some tests accordingly to the changes.
