+++ Source/classifiers.py
@@ -1,3 +1,4 @@
+"""This module is not in used yet! you can ignore it for now"""
 from torch.utils.data import dataset
 import torch.nn as nn
 from Source.pipelines import Model_Pipeline
