+++ Tests/conftest.py
@@ -2,7 +2,7 @@ import pytest
 import numpy as np
 import mne
 from Source.pipelines import Data_Pipeline
-from Source.recording import Recording
+from Source.data_manager import Recording
 from pathlib import Path
 from Source import pipelines
 from _pytest.monkeypatch import monkeypatch
