+++ requirments.txt
@@ -1,0 +1,8 @@
+pandas==2.0.0
+mne==1.3.1
+scikit_learn==1.2.2
+torch==2.0.0
+pytest==7.3.0
+hmmlearn==0.2.8
+jupyterlab==3.6.3
+jupyter==1.0.0
